hint-service/requirements.txt
fastapi==0.115.0
uvicorn==0.30.6
sentence-transformers==3.0.1
transformers==4.44.2
torch==2.4.0
accelerate==0.33.0
supabase==2.7.2
python-dotenv==1.0.1
pydantic==2.8.2
 